Mongo Db database is added to the setting
Using djongo connector django is connected to the mongodb database. Also should edit `site-packages/pymongo/mongo_client.py line-90`  this file in the virtual env for system for the connection to remote database. The HOST variable should be changed in the file which default value is local host but should be changed to the same value as of settings.py > Database > default >Host