Creating custom exception Handler.
When the user is not authenticated before only detail was on response now it will have
detail,
message,
status_code.
The detail is of no use now as it is same of message. But I didn't get any idea to delete the detail so I am including it in the response. In future we can delete that.